set selection mode to 'exclusive'
This will cause the selection to stop before the cursor. Otherwise the
character below the cursor is included.

This allows to use V$ S( to wrap the selection with brackets. Otherwise
the new line character would be included in the selection and so also
get surrounded causing the closing bracket to be placed on the next
line.